src/requirements.txt
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
requests==2.27.1
smmap==5.0.0
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
